Rearrange council a little
